requirements.txt
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-storage==1.13.0

python-gitlab==1.6
